requirements.txt
langchain 
openai 
chromadb 
pandas 
streamlit
python-dotenv
langchain-google-genai 
langchain-community 
langchain-chroma
google-generativeai>=0.8.3
